functions/ZohoAuth/requirements.txt
crowdstrike-foundry-function==0.5.3
